+++ source/benoit/benoit/app/handle_keys.rs
@@ -21,12 +21,11 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::benoit::PRECISION;
+use crate::benoit::{PRECISION};
 use crate::benoit::app::App;
-use crate::benoit::factorisation::Factorisation;
 use crate::benoit::fractal::Fractal;
 use crate::benoit::palette::Palette;
-use crate::benoit::render::{FactoriserFunction, IteratorFunction, PaletteFunction, RowRenderer};
+use crate::benoit::render::{IteratorFunction, RowRenderer};
 use crate::benoit::rendering::Rendering;
 
 extern crate rug;
@@ -35,17 +34,19 @@ extern crate sdl2;
 use rug::Float;
 use sdl2::keyboard::Scancode;
 
+pub const MIN_COLOUR_RANGE: f32 = 2.0;
+
 impl App {
+	#[must_use]
 	pub fn handle_keys(&mut self, scan_code: Scancode) -> bool {
 		match scan_code {
 			Scancode::C      => self.do_render = true,
 			Scancode::Escape => return true,
 			Scancode::F1     => self.do_textual_feedback = !self.do_textual_feedback,
 			Scancode::LAlt   => (self.fractal, self.multibrot_exponent, self.iterator_function) = cycle_fractal(self.fractal, -0x1),
-			Scancode::LCtrl  => (self.factorisation, self.factoriser) = toggle_smooth(self.factorisation),
-			Scancode::Left   => (self.palette, self.palette_function) = cycle_palette(self.palette, -0x1),
+			Scancode::Left   => self.palette = cycle_palette(self.palette, -0x1),
 			Scancode::RAlt   => (self.fractal, self.multibrot_exponent, self.iterator_function) = cycle_fractal(self.fractal, 0x1),
-			Scancode::Right  => (self.palette, self.palette_function) = cycle_palette(self.palette, 0x1),
+			Scancode::Right  => self.palette = cycle_palette(self.palette, 0x1),
 			Scancode::Tab    => (self.rendering, self.row_renderer) = toggle_julia(self.rendering),
 			Scancode::Z      => eprintln!("c = {}{:+}i, {}x @ {} iter. (range.: {:.3})", &self.centre_real, &self.centre_imag, &self.zoom, self.max_iter_count, self.colour_range),
 			_                => {},
@@ -63,7 +64,7 @@ impl App {
 
 		self.colour_range = match scan_code {
 			Scancode::Up   => self.colour_range * COLOUR_RANGE_FACTOR,
-			Scancode::Down => (self.colour_range / COLOUR_RANGE_FACTOR).max(2.0),
+			Scancode::Down => (self.colour_range / COLOUR_RANGE_FACTOR).max(MIN_COLOUR_RANGE),
 			_              => self.colour_range,
 		};
 
@@ -125,24 +126,10 @@ fn toggle_julia(rendering: Rendering) -> (Rendering, RowRenderer) {
 	return (rendering, row_renderer);
 }
 
-fn toggle_smooth(factorisation: Factorisation) -> (Factorisation, FactoriserFunction) {
-	let factorisation = factorisation.cycle();
-
-	let factoriser = factorisation.get_factoriser();
-
-	match factorisation {
-		Factorisation::Smooth  => eprintln!("disabled smoothing"),
-		Factorisation::Stepped => eprintln!("enabled smoothing"),
-	};
-
-	return (factorisation, factoriser);
-}
-
-fn cycle_palette(palette: Palette, direction: i8) -> (Palette, PaletteFunction) {
-	let palette = palette + direction;
-	let palette_function = palette.get_function();
+fn cycle_palette(palette: Palette, direction: i8) -> Palette {
+	let palette = palette.cycle(direction);
 
 	eprintln!("using palette \"{}\"", palette.get_name());
 
-	return (palette, palette_function);
+	return palette;
 }