+++ source/benoit/benoit/palette/paint/greyscale.rs
@@ -21,23 +21,17 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::benoit::app::App;
-use crate::benoit::render::colour_data::ColourData;
+// Original palette function after the Rust-
+// rewrite (version 10).
 
-extern crate rayon;
+pub fn greyscale(factor: f32) -> (f32, f32, f32) {
+	let factor = factor % 1.0;
 
-use rayon::prelude::*;
-use std::sync::Arc;
+	let factor = (if factor >= 1.0 / 2.0 {
+		1.0 - factor
+	} else {
+		factor
+	}) * 2.0;
 
-impl App {
-	pub fn colour(&self, buffer: &mut [u8], multibrot_exponent: f32, max_iter_count: u32, iter_count_buffer: &[u32], square_dist_buffer: &[f32]) {
-		let data = Arc::new(ColourData::new(buffer, self.canvas_width, multibrot_exponent, max_iter_count, self.colour_range, iter_count_buffer, square_dist_buffer));
-
-		let factoriser       = self.factoriser;
-		let palette_function = self.palette_function;
-
-		(0x0..self.canvas_height).into_par_iter().for_each(|row| {
-			App::colour_row(data.clone(), row as u32, factoriser, palette_function);
-		});
-	}
+	return (factor, factor, factor);
 }